requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
spacy==3.8.5
openai==1.35.13
python-dotenv==1.0.1
psycopg2-binary==2.9.9
networkx
matplotlib==3.9.2
